SF-344 Auth0 Sys Admin Users tab
* remove add user
* remove view/edit user
* fix responsive now that drawer is never open
* move projects component into SA folder
* rename delete dialog buttons
* fix unit tests
* Convert some `mat` to `mdc`
* center content on large viewports
