+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/sa-delete-dialog.component.html
@@ -1,6 +1,20 @@
-<h3 mat-dialog-title>Delete Confirmation</h3>
-<mat-dialog-content> <p>Are you sure you want to delete?</p> </mat-dialog-content>
-<mat-dialog-actions align="end">
-  <button mat-raised-button (click)="closeDialog()" id="confirm-button-no">No</button>
-  <button mat-raised-button class="mat-primary" (click)="confirmDialog()" id="confirm-button-yes">Yes</button>
-</mat-dialog-actions>
+<mdc-dialog>
+  <mdc-dialog-container>
+    <mdc-dialog-surface>
+      <mdc-dialog-title>Delete Confirmation</mdc-dialog-title>
+      <mdc-dialog-content>
+        <div fxLayout="row" fxLayoutAlign="center center" fxLayoutGap="10px">
+          <app-avatar [user]="data.user" size="32" [round]="false"> </app-avatar>
+          <div>
+            <strong>{{ data.user.name }}</strong> <br />
+            {{ data.user.email }}
+          </div>
+        </div>
+      </mdc-dialog-content>
+      <mdc-dialog-actions>
+        <button mdcDialogButton mdcDialogAction="close" id="confirm-button-no">Cancel</button>
+        <button mdcDialogButton default mdcDialogAction="confirmed" id="confirm-button-yes">Delete</button>
+      </mdc-dialog-actions>
+    </mdc-dialog-surface>
+  </mdc-dialog-container>
+</mdc-dialog>
