+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/sa-delete-dialog.component.spec.ts
@@ -1,44 +1,93 @@
-import { CUSTOM_ELEMENTS_SCHEMA, DebugElement, NO_ERRORS_SCHEMA } from '@angular/core';
+import { MDC_DIALOG_DATA, MdcDialog, MdcDialogConfig, MdcDialogRef } from '@angular-mdc/web';
+import { CommonModule } from '@angular/common';
+import {
+  Component,
+  CUSTOM_ELEMENTS_SCHEMA,
+  DebugElement,
+  Directive,
+  NgModule,
+  ViewChild,
+  ViewContainerRef
+} from '@angular/core';
 import { ComponentFixture, fakeAsync, flush, TestBed } from '@angular/core/testing';
-import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material';
 import { By } from '@angular/platform-browser';
-import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { instance, mock } from 'ts-mockito';
-
 import { UICommonModule } from '../ui-common.module';
-import { UserService } from '../user.service';
-import { SaDeleteDialogComponent } from './sa-delete-dialog.component';
+import { SaDeleteDialogComponent, SaDeleteUserDialogData } from './sa-delete-dialog.component';
 
-class TestDeleteDialogComponent {
-  dialog: MatDialog;
+describe('DeleteDialogComponent', () => {
+  it('Confirm Delete button call', fakeAsync(() => {
+    const env = new TestEnvironment();
+    env.clickElement(env.deleteButton);
+    flush();
+    expect(env.afterCloseCallback).toHaveBeenCalledWith('confirmed');
+  }));
+
+  it('Confirm Cancel button call', fakeAsync(() => {
+    const env = new TestEnvironment();
+    env.clickElement(env.cancelButton);
+    flush();
+    expect(env.afterCloseCallback).toHaveBeenCalledWith('close');
+  }));
+});
+
+@Directive({
+  // ts lint complains that a directive should be used as an attribute
+  // tslint:disable-next-line:directive-selector
+  selector: 'viewContainerDirective'
+})
+class ViewContainerDirective {
+  constructor(public viewContainerRef: ViewContainerRef) {}
+}
+
+@Component({
+  selector: 'app-view-container',
+  template: '<viewContainerDirective></viewContainerDirective>'
+})
+class ChildViewContainerComponent {
+  @ViewChild(ViewContainerDirective) viewContainer: ViewContainerDirective;
+
+  get childViewContainer(): ViewContainerRef {
+    return this.viewContainer.viewContainerRef;
+  }
+}
+
+@NgModule({
+  imports: [CommonModule, UICommonModule],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+  declarations: [ViewContainerDirective, ChildViewContainerComponent, SaDeleteDialogComponent],
+  exports: [ViewContainerDirective, ChildViewContainerComponent, SaDeleteDialogComponent],
+  entryComponents: [ChildViewContainerComponent, SaDeleteDialogComponent]
+})
+class TestModule {}
+
+class TestEnvironment {
   component: SaDeleteDialogComponent;
-  fixture: ComponentFixture<SaDeleteDialogComponent>;
-
-  mockedUserService: UserService;
+  fixture: ComponentFixture<ChildViewContainerComponent>;
+  dialogRef: MdcDialogRef<SaDeleteDialogComponent>;
+  afterCloseCallback: jasmine.Spy;
 
   constructor() {
-    this.mockedUserService = mock(UserService);
-
     TestBed.configureTestingModule({
-      imports: [NoopAnimationsModule, UICommonModule],
-      declarations: [SaDeleteDialogComponent],
-      schemas: [CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA],
-      providers: [
-        { provide: UserService, useFactory: () => instance(this.mockedUserService) },
-        { provide: MAT_DIALOG_DATA },
-        { provide: MatDialogRef, useValue: {} }
-      ]
+      imports: [TestModule]
     });
-
-    this.fixture = TestBed.createComponent(SaDeleteDialogComponent);
-    this.component = this.fixture.componentInstance;
+    this.fixture = TestBed.createComponent(ChildViewContainerComponent);
+    const viewContainerRef = this.fixture.componentInstance.childViewContainer;
+    const config: MdcDialogConfig<SaDeleteUserDialogData> = {
+      viewContainerRef,
+      data: { user: { id: '', type: '', name: 'Billy T James', email: 'w.j.t.w.taitoko@example.com' } }
+    };
+    this.dialogRef = TestBed.get(MdcDialog).open(SaDeleteDialogComponent, config);
+    this.afterCloseCallback = jasmine.createSpy('afterClose callback');
+    this.dialogRef.afterClosed().subscribe(this.afterCloseCallback);
+    this.component = this.dialogRef.componentInstance;
+    this.fixture.detectChanges();
   }
 
-  get noButton(): DebugElement {
+  get cancelButton(): DebugElement {
     return this.fixture.debugElement.query(By.css('#confirm-button-no'));
   }
 
-  get yesButton(): DebugElement {
+  get deleteButton(): DebugElement {
     return this.fixture.debugElement.query(By.css('#confirm-button-yes'));
   }
 
@@ -52,21 +101,3 @@ class TestDeleteDialogComponent {
     flush();
   }
 }
-
-describe('DeleteDialogComponent', () => {
-  it('Confirm Yes button call', fakeAsync(() => {
-    const env = new TestDeleteDialogComponent();
-    spyOn(env.component, 'confirmDialog');
-    env.clickElement(env.yesButton);
-
-    expect(env.component.confirmDialog).toHaveBeenCalled();
-  }));
-
-  it('Confirm No button call', fakeAsync(() => {
-    const env = new TestDeleteDialogComponent();
-    spyOn(env.component, 'closeDialog');
-    env.clickElement(env.noButton);
-
-    expect(env.component.closeDialog).toHaveBeenCalled();
-  }));
-});
