+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/sa-delete-dialog.component.ts
@@ -1,20 +1,14 @@
+import { MDC_DIALOG_DATA } from '@angular-mdc/web';
 import { Component, Inject } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
+import { User } from 'xforge-common/models/user';
+
+export interface SaDeleteUserDialogData {
+  user: User;
+}
 
 @Component({
   templateUrl: './sa-delete-dialog.component.html'
 })
 export class SaDeleteDialogComponent {
-  constructor(
-    private readonly dialogRef: MatDialogRef<SaDeleteDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) private data: any
-  ) {}
-
-  confirmDialog(): void {
-    this.dialogRef.close('confirmed');
-  }
-
-  closeDialog(): void {
-    this.dialogRef.close('notconfirmed');
-  }
+  constructor(@Inject(MDC_DIALOG_DATA) public data: SaDeleteUserDialogData) {}
 }
