--- src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/sa-user-entry.component.html
@@ -1,69 +1,0 @@
-<mat-card>
-  <mat-card-title id="title" class="account-title"> {{ headerTitle }} </mat-card-title>
-  <mat-card-content fxLayout="column">
-    <div fxLayout="column" fxLayoutAlign="center center">
-      <div class="fields-container">
-        <form
-          [formGroup]="accountUserForm"
-          #accountUserFormReset="ngForm"
-          fxLayout="column"
-          novalidate
-          autocomplete="off"
-        >
-          <mat-form-field>
-            <input id="full-name" matInput placeholder="Name" formControlName="fullName" />
-            <mat-error *ngIf="fullName.errors?.required && isSubmitted">Name is required</mat-error>
-          </mat-form-field>
-
-          <mat-form-field>
-            <input id="email" matInput placeholder="Email Address" formControlName="email" />
-            <mat-error *ngIf="email.errors?.required && isSubmitted"> Email is required </mat-error>
-            <mat-error *ngIf="(email.errors?.email || email.errors?.pattern) && isSubmitted">
-              Email is invalid
-            </mat-error>
-          </mat-form-field>
-
-          <mat-form-field>
-            <mat-select id="user-role" placeholder="System role" formControlName="role">
-              <mat-option *ngFor="let item of roleList" value="{{ item.id }}">{{ item.value }}</mat-option>
-            </mat-select>
-            <mat-error *ngIf="role.errors?.required && isSubmitted"> System Role is required </mat-error>
-          </mat-form-field>
-          <div *ngIf="showActivateDeActivatePanel">
-            <p id="last-login" mat-line>Last login {{ userLastLoginDate }}</p>
-            <p id="date-created" mat-line>User created {{ userCreatedDate }}</p>
-            <section class="example-section">
-              <mat-slide-toggle class="example-margin" formControlName="activateStatus" (change)="onChange($event)">
-                {{ userActivateDeactive }}
-              </mat-slide-toggle>
-            </section>
-          </div>
-          <mat-card-actions align="end">
-            <button
-              mat-raised-button
-              id="add-button"
-              color="primary"
-              *ngIf="btnUserAdd"
-              type="submit"
-              class="add-button"
-              (click)="onUserAdd()"
-            >
-              Add
-            </button>
-            <button
-              mat-raised-button
-              id="update-button"
-              color="primary"
-              *ngIf="btnUserUpdate"
-              type="button"
-              class="update-button"
-              (click)="onUpdate()"
-            >
-              Update
-            </button>
-          </mat-card-actions>
-        </form>
-      </div>
-    </div>
-  </mat-card-content>
-</mat-card>
