+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/sa-users.component.scss
@@ -1,3 +1,7 @@
+.users-controls {
+  margin: 32px 0;
+}
+
 .mat-column-avatar {
   width: 65px !important;
 }
@@ -17,11 +21,6 @@ mat-form-field {
 
 #no-users-label {
   padding-top: 14px;
-}
-
-.loading-spinner {
-  align-self: center;
-  margin: 30px 0px;
 }
 
 .remove-cell {
