+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/sa-users.component.ts
@@ -1,7 +1,7 @@
-import { Component, HostBinding, OnInit } from '@angular/core';
-import { MatDialog, MatDialogRef } from '@angular/material';
+import { MdcDialog, MdcDialogConfig, MdcDialogRef } from '@angular-mdc/web';
+import { Component, HostBinding, OnDestroy, OnInit } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
-
+import { NoticeService } from 'xforge-common/notice.service';
 import { GetAllParameters } from '../json-api.service';
 import { Project } from '../models/project';
 import { ProjectUser } from '../models/project-user';
@@ -9,7 +9,7 @@ import { User } from '../models/user';
 import { SubscriptionDisposable } from '../subscription-disposable';
 import { UserService } from '../user.service';
 import { nameof } from '../utils';
-import { SaDeleteDialogComponent } from './sa-delete-dialog.component';
+import { SaDeleteDialogComponent, SaDeleteUserDialogData } from './sa-delete-dialog.component';
 
 interface Row {
   readonly user: User;
@@ -21,29 +21,34 @@ interface Row {
   templateUrl: './sa-users.component.html',
   styleUrls: ['./sa-users.component.scss']
 })
-export class SaUsersComponent extends SubscriptionDisposable implements OnInit {
+export class SaUsersComponent extends SubscriptionDisposable implements OnInit, OnDestroy {
   @HostBinding('class') classes = 'flex-column';
 
   length: number = 0;
   pageIndex: number = 0;
   pageSize: number = 50;
-  showAddPanel: boolean = false;
-  showEditPanel: boolean = false;
 
   userRows: Row[];
 
-  private userId: string;
-  private dialogRef: MatDialogRef<SaDeleteDialogComponent, string>;
-
+  private dialogRef: MdcDialogRef<SaDeleteDialogComponent>;
   private readonly searchTerm$: BehaviorSubject<string>;
   private readonly parameters$: BehaviorSubject<GetAllParameters<User>>;
   private readonly reload$: BehaviorSubject<void>;
 
-  constructor(private readonly dialog: MatDialog, private readonly userService: UserService) {
+  constructor(
+    private readonly dialog: MdcDialog,
+    private readonly noticeService: NoticeService,
+    private readonly userService: UserService
+  ) {
     super();
+    this.noticeService.loadingStarted();
     this.searchTerm$ = new BehaviorSubject<string>('');
     this.parameters$ = new BehaviorSubject<GetAllParameters<User>>(this.getParameters());
     this.reload$ = new BehaviorSubject<void>(null);
+  }
+
+  get isLoading(): boolean {
+    return this.userRows == null;
   }
 
   ngOnInit() {
@@ -51,6 +56,7 @@ export class SaUsersComponent extends SubscriptionDisposable implements OnInit {
     this.subscribe(
       this.userService.onlineSearch(this.searchTerm$, this.parameters$, this.reload$, include),
       searchResults => {
+        this.noticeService.loadingStarted();
         if (searchResults && searchResults.data) {
           this.userRows = searchResults.data.map(user => {
             const projects = searchResults
@@ -60,12 +66,14 @@ export class SaUsersComponent extends SubscriptionDisposable implements OnInit {
           });
           this.length = searchResults.totalPagedCount;
         }
+        this.noticeService.loadingFinished();
       }
     );
   }
 
-  get isLoading(): boolean {
-    return this.userRows == null;
+  ngOnDestroy(): void {
+    super.ngOnDestroy();
+    this.noticeService.loadingFinished();
   }
 
   updateSearchTerm(term: string): void {
@@ -78,40 +86,18 @@ export class SaUsersComponent extends SubscriptionDisposable implements OnInit {
     this.parameters$.next(this.getParameters());
   }
 
-  addUser(): void {
-    this.showEditPanel = false;
-    this.showAddPanel = !this.showAddPanel;
-    this.userId = '';
-  }
-
-  editUser(userId: string): void {
-    if (this.showAddPanel || this.userId !== userId) {
-      this.showEditPanel = true;
-      this.showAddPanel = false;
-    } else {
-      this.showEditPanel = !this.showEditPanel;
-    }
-    this.userId = userId;
-  }
-
-  removeUser(userId: string): void {
-    this.dialogRef = this.dialog.open(SaDeleteDialogComponent, {
-      width: '350px',
-      height: '200px',
-      disableClose: true
-    });
+  removeUser(user: User): void {
+    const dialogConfig: MdcDialogConfig<SaDeleteUserDialogData> = {
+      data: {
+        user
+      }
+    };
+    this.dialogRef = this.dialog.open(SaDeleteDialogComponent, dialogConfig);
     this.dialogRef.afterClosed().subscribe(confirmation => {
       if (confirmation.toLowerCase() === 'confirmed') {
-        this.showEditPanel = false;
-        this.deleteUser(userId);
+        this.deleteUser(user.id);
       }
     });
-  }
-
-  outputUserList(): void {
-    this.showEditPanel = false;
-    this.showAddPanel = false;
-    this.reload$.next(null);
   }
 
   private async deleteUser(userId: string) {
