+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/system-administration.component.scss
@@ -1,0 +1,8 @@
+.body-content {
+  max-width: 800px;
+  margin: 0 auto;
+}
+
+app-sa-users {
+  width: 100%;
+}
