+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/system-administration/system-administration.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-system-administration',
-  templateUrl: './system-administration.component.html'
+  templateUrl: './system-administration.component.html',
+  styleUrls: ['./system-administration.component.scss']
 })
 export class SystemAdministrationComponent {}
