+++ src/SIL.XForge.Scripture/ClientApp/src/xforge-common/ui-common.module.ts
@@ -26,27 +26,12 @@ import { NgModule } from '@angular/core';
 import { BREAKPOINT, FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import {
-  MatButtonModule,
-  MatButtonToggleModule,
-  MatCardModule,
-  MatCheckboxModule,
-  MatDatepickerModule,
-  MatDialogModule,
-  MatDividerModule,
   MatFormFieldModule,
-  MatGridListModule,
-  MatIconModule,
-  MatInputModule,
-  MatNativeDateModule,
   MatOptionModule,
   MatPaginatorModule,
-  MatProgressBarModule,
   MatProgressSpinnerModule,
   MatSelectModule,
-  MatSlideToggleModule,
-  MatSnackBarModule,
-  MatTableModule,
-  MatTabsModule
+  MatTableModule
 } from '@angular/material';
 import { ChartsModule } from 'ng2-charts';
 
@@ -56,27 +41,12 @@ const modules = [
   ChartsModule,
   FlexLayoutModule,
   FormsModule,
-  MatButtonModule,
-  MatButtonToggleModule,
-  MatCardModule,
-  MatCheckboxModule,
-  MatDatepickerModule,
-  MatDialogModule,
-  MatDividerModule,
   MatFormFieldModule,
-  MatGridListModule,
-  MatIconModule,
-  MatInputModule,
-  MatNativeDateModule,
   MatOptionModule,
   MatPaginatorModule,
-  MatProgressBarModule,
   MatProgressSpinnerModule,
   MatSelectModule,
-  MatSlideToggleModule,
-  MatSnackBarModule,
   MatTableModule,
-  MatTabsModule,
   MdcButtonModule,
   MdcCardModule,
   MdcCheckboxModule,
